add after_update_commit broadcasting to ExchangeRate model (#5)

+++ app/jobs/delete_dummy_job.rb
@@ -2,6 +2,6 @@ class DeleteDummyJob < ApplicationJob
   queue_as :default
 
   def perform(dummy)
-    dummy.destroy
+    dummy.destroy if dummy.persisted?
   end
 end
